move betav1 test
